Reframe UTXO attractor examples for accepted harness
